Add experimental generation of graph data
This isn't very useful on its own, but combined with some d3 code it can
produce a graph showing the relationships between every package (see
https://bl.ocks.org/bobwhitelock/raw/38f20becc90ad07e552d1a8551e33d46/).
Which still isn't very useful, but could possibly be adapted to be.

To get the graph to display also had to remove
`eskimoblood/elm-simplex-noise` package from the data, as this depends
on the `elm-community/elm-random-extra` package which is not in the
latest packages list (as it has not been updated for 0.18); this then
caused everything to blow up.