Have request-packages check URLs in generic way
Rather than having to repeatedly catch same exception, check each URL in
turn and raise if file not successfully found by the time the last is
requested. This will make it easier to add more URLs to the list to
check and have these handled in the same way.

Script will now currently error when run for all packages, as there are
still some packages which will always give a 404 for all URLs, for a
couple of reasons to be addressed shortly.