Display package dependents as list within column
The dependent packages for each package are now displayed so the number
of dependents can always be seen, while the names of these dependent
packages, if any, are displayed as a list hidden within a `details`
element. The dependent names are shown as buttons to set the query in
the same way as the dependency names.