Remove Dependencies type
Now that the `request-packages` script should only include packages
where we can retrieve an `elm-package.json` from GitHub, we no longer
need to handle the situation where this fails. Therefore the
`Dependencies` type, which modelled this possibility, can be removed and
just a list used in its place, which simplifies things in various
places.