Load packages on page load if already authed
On page load, check if we already have a (probably) valid session with
GitHub, and if so retrieve the access token and include this in the
program flags. This is then decoded along with the initial packages, and
if present we can skip ahead the `PackagesData` state and start loading
the GraphQL data from GitHub, without needing to click a button to
re-authenticate first.